+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog #
 
+## Unreleased ##
+ * Use username as a backup `displayName`
+
 ## 0.21 ##
  * Create FIDO errors list when needed unless it already exists
  * Clear FIDO errors list before each registration or authentication request